[STD-33] Fixing tests - Creating tests for event controller - Organizing CMakeLists.txt

+++ src/ui/event/ListenerWrapper.cpp
@@ -12,7 +12,7 @@ namespace Catalyst {
         listeners.remove(listener);
     }
 
-    int ListenerWrapper::listenersSize() {
+    size_t ListenerWrapper::listenersSize() {
         return listeners.getLength();
     }
